GPG-460:Refactoring some models
