+++ GenderPayGap.WebUI/Models/Report/ReportOverviewSectionsViewModel.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using GovUkDesignSystem;
 using Microsoft.AspNetCore.Html;
 
-namespace GenderPayGap.WebUI.Views.Components.ReportOverview
+namespace GenderPayGap.WebUI.Models.Report
 {
     public class ReportOverviewSectionsViewModel
     {
