+++ GenderPayGap.WebUI/Views/Components/ReportOverview/ReportOverviewSections.cshtml
@@ -1,5 +1,5 @@
-﻿@using GenderPayGap.WebUI.Views.Components.ReportOverview
-@model GenderPayGap.WebUI.Views.Components.ReportOverview.ReportOverviewSectionsViewModel
+﻿@using GenderPayGap.WebUI.Models.Report
+@model GenderPayGap.WebUI.Models.Report.ReportOverviewSectionsViewModel
 
 <h3 class="govuk-heading-s">@(Model.Title)</h3>
 <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
