+++ GenderPayGap.WebUI/Views/ReportOverview/ReportOverview.cshtml
@@ -1,9 +1,9 @@
 @using GenderPayGap.Core.Helpers
-@using GenderPayGap.WebUI.Views.Components.ReportOverview
 @using GovUkDesignSystem
 @using GovUkDesignSystem.GovUkDesignSystemComponents
 @using System.ComponentModel.DataAnnotations
 @using GenderPayGap.Core
+@using GenderPayGap.WebUI.Models.Report
 @using GenderPayGap.WebUI.Views.Components.StatusMessages
 @using GovUkDesignSystem.GovUkDesignSystemComponents.SubComponents
 @model GenderPayGap.WebUI.Models.Report.ReportOverviewViewModel
